Rename to "translator_plugins" per  samsonasik's request
